// src/interfaces/Student.ts
export interface Student {
  address: string;
  dateOfBirth: string;
  enrollmentDate: string;
  examCandidateNumber: string;
  firstName: string;
  house: string;
  lastName: string;
  photoConsent: boolean;
  previousSchool: number;
  studentId: number;
  subjects: number[];
  tripConsent: boolean;
  uci: string;
  tutor: number;
  yearGroup: number;
  schoolId: number;
} 